frontend

+++ frontend/src/components/admin/Blogs/BlogPage.jsx
@@ -1,31 +1,51 @@
-// src/pages/AdminBlogPage.js
-
-import React, { useState, useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
+import React, { useState } from "react";
 import {
   Container,
   Row,
   Col,
   Button,
-  Form,
   Spinner,
   Alert,
+  Form,
 } from "react-bootstrap";
-import { fetchBlogs, deleteBlog } from "../../../redux/actions/blogActions";
-import AdminBlogCard from "./AdminBlogCard";
+import { useDispatch } from "react-redux";
+import { deleteBlog } from "../../../redux/actions/blogActions";
+import BlogCard from "../../BlogCard/BlogCard";
 import { Link } from "react-router-dom";
+import useFetch from "../../../hooks/useFetch";
+import Swal from "sweetalert2";
+import { BASE_URL } from "../../../utils/config";
 
 const AdminBlogPage = () => {
   const dispatch = useDispatch();
-  const { blogs, loading, error } = useSelector((state) => state.blog);
   const [searchTerm, setSearchTerm] = useState("");
+  const { data: blogs, loading, error } = useFetch("blogs"); // Adjust URL as needed
+  const [selectedBlog, setSelectedBlog] = useState(null);
 
-  useEffect(() => {
-    dispatch(fetchBlogs());
-  }, [dispatch]);
+  const handleDelete = async (id) => {
+    try {
+      const result = await Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!",
+      });
 
-  const handleDelete = (id) => {
-    dispatch(deleteBlog(id));
+      if (result.isConfirmed) {
+        await dispatch(deleteBlog(id));
+        Swal.fire("Deleted!", "The blog has been deleted.", "success");
+      }
+    } catch (err) {
+      Swal.fire(
+        "Error!",
+        "There was an error deleting the blog. Please try again later.",
+        "error"
+      );
+      console.error("Error deleting blog:", err);
+    }
   };
 
   const filteredBlogs = Array.isArray(blogs)
@@ -37,6 +57,47 @@ const AdminBlogPage = () => {
   const handleSearch = (e) => {
     setSearchTerm(e.target.value);
   };
+
+  const renderBlogs = () => {
+    if (error) {
+      return (
+        <Alert variant="danger" className="mt-4">
+          {error}
+        </Alert>
+      );
+    }
+
+    if (filteredBlogs.length === 0) {
+      return (
+        <Alert variant="warning" className="mt-4">
+          No blogs found.
+        </Alert>
+      );
+    }
+
+    return (
+      <Container>
+        <Row>
+          {filteredBlogs.map((blog) => (
+            <Col lg="3" sm="6" key={blog._id} md={4} className="mb-4">
+              <BlogCard blog={blog} onDelete={handleDelete} isAdmin={true} />
+            </Col>
+          ))}
+        </Row>
+      </Container>
+    );
+  };
+
+  if (loading) {
+    return (
+      <Container
+        className="d-flex justify-content-center align-items-center"
+        style={{ height: "100vh" }}
+      >
+        <Spinner animation="border" variant="primary" />
+      </Container>
+    );
+  }
 
   return (
     <Container className="mt-4">
@@ -62,21 +123,7 @@ const AdminBlogPage = () => {
           </Form>
         </Col>
       </Row>
-      <Row>
-        {loading ? (
-          <Spinner animation="border" variant="primary" />
-        ) : error ? (
-          <Alert variant="danger">{error}</Alert>
-        ) : filteredBlogs.length === 0 ? (
-          <Alert variant="info">No blogs found.</Alert>
-        ) : (
-          filteredBlogs.map((blog) => (
-            <Col key={blog._id} md={4} className="mb-4">
-              <AdminBlogCard blog={blog} onDelete={handleDelete} />
-            </Col>
-          ))
-        )}
-      </Row>
+      {renderBlogs()}
     </Container>
   );
 };
